+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 1. Fix escaped quotes in formula
 2. .ojson is valid extension for AA
+3. languageId for Oscript is oscript instead of ojson
 
 ## v0.2.3